+++ src/Core/Domain/index.ts
@@ -3,3 +3,4 @@ export * from "./ServiceModel";
 export * from "./ResourceModel";
 export * from "./ServiceInstanceModel";
 export * from "./ProjectModel";
+export * as Query from "./Query";