+++ src/UI/ServiceInventory/InstanceRow.tsx
@@ -1,4 +1,4 @@
-import { InstanceForResources, Row } from "@/Core";
+import { Row, ServiceInstanceIdentifier } from "@/Core";
 import React, { useState } from "react";
 import { Tbody, Tr, Td, ExpandableRowContent } from "@patternfly/react-table";
 import { words } from "@/UI";
@@ -19,6 +19,7 @@ import {
   InfoCircleIcon,
   ListIcon,
 } from "@patternfly/react-icons";
+import { DeploymentProgressPresenter } from "./Presenters";
 
 interface Props {
   row: Row;
@@ -28,7 +29,7 @@ interface Props {
   numberOfColumns: number;
   actions: React.ReactElement | null;
   state: React.ReactElement | null;
-  instanceForResources: InstanceForResources;
+  serviceInstanceIdentifier: ServiceInstanceIdentifier;
 }
 
 export const InstanceRow: React.FC<Props> = ({
@@ -39,9 +40,10 @@ export const InstanceRow: React.FC<Props> = ({
   numberOfColumns,
   actions,
   state,
-  instanceForResources,
+  serviceInstanceIdentifier,
 }) => {
   const [activeTab, setActiveTab] = useState<TabKey>("Status");
+  const deploymentProgressPresenter = new DeploymentProgressPresenter();
   const attributesOnClick = () => {
     if (!isExpanded) {
       onToggle();
@@ -69,6 +71,11 @@ export const InstanceRow: React.FC<Props> = ({
               onClick={attributesOnClick}
             />
           </a>
+        </Td>
+        <Td dataLabel={words("inventory.collumn.deploymentProgress")}>
+          {deploymentProgressPresenter.getDeploymentProgressBar(
+            row.deploymentProgress
+          )}
         </Td>
         <Td dataLabel={words("inventory.column.createdAt")}>
           <DateWithTooltip date={row.createdAt} />
@@ -102,7 +109,7 @@ export const InstanceRow: React.FC<Props> = ({
                 icon={<ListIcon />}
               />
               <ResourcesView
-                instance={instanceForResources}
+                qualifier={serviceInstanceIdentifier}
                 title={words("inventory.tabs.resources")}
                 icon={<AutomationIcon />}
               />