+++ src/UI/index.ts
@@ -1,3 +1,4 @@
 export * from "./ServiceInventory";
 export * from "./words";
 export * from "./Store";
+export * from "./ServicesContext";